% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BinSpp}
\alias{BinSpp}
\title{A List with two Stacks each for an hypothetic species}
\format{A list with two stacks of the distribution of two ypothetical species:
\describe{
  \item{SeciesA}{A stack with eight time slices for species A}
  \item{SpeciesB}{A stack with eight time slices for species b}
}}
\usage{
BinSpp
}
\description{
A dataset containing two stacks in a list, each contains eight rasters, each represents
a timeslice of the distribution of the species
}
\keyword{datasets}
