% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Cost}
\alias{Cost}
\title{A raster with a simulated cost for our hypothetical species}
\format{Raster
\describe{
  \item{Cost}{simulated cost layer}
}}
\usage{
Cost
}
\description{
This raster has areas with cost Zero (protected areas), unavailable cells
(cities or agricultural areas) and cells with cost.
}
\keyword{datasets}
