% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{GetQuadFlow}
\alias{GetQuadFlow}
\title{Imports the AMPL result back to R to get the timeslice flow}
\usage{
GetQuadFlow(Stacklist, AmplFile, Species, plot = TRUE, gif = FALSE)
}
\arguments{
\item{Stacklist}{The original stacklist used in function MultiSppQuad}

\item{AmplFile}{The path to the file given by the AMPL model}

\item{Species}{an interger number of the species to be ploted, the
data frame is made for all the species}

\item{plot}{logical, wether to plot or not the stack, defaults to TRUE}

\item{gif}{logical, wether to make a gif or not the stack, defaults to FALSE}
}
\value{
a list with a stack with a stack with the
index for each species, and a data frame with the results
}
\description{
From the result gotten from the AMPL model, it will develop a stack with the
flow for each species on each time slice, and a data frame with the results
}
\examples{
\dontrun{
#Load the data
data("BinSpp")
#Use the function with the
Index <- GetQuadFlow(Stacklist = BinSpp, AmplFile = "https://raw.githubusercontent.com/derek-corcoran-barrios/QuadraticCostScripts/master/TwoY.txt", Species = 1, plot = FALSE, gif = TRUE)
}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
