% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{GetQuadIndex}
\alias{GetQuadIndex}
\title{Imports the AMPL result back to R}
\usage{
GetQuadIndex(Stacklist, AmplFile, plot = TRUE)
}
\arguments{
\item{Stacklist}{The original stacklist used in function MultiSppQuad}

\item{AmplFile}{The path to the file given by the AMPL model}

\item{plot}{logical, wether to plot or not the stack, defaults to TRUE}
}
\value{
a list with a stack with a stack with the
index for each species, and a data frame with the results
}
\description{
From the result gotten from the AMPL model, it will develop a stack with the
index for each species, and a data frame with the results
}
\examples{
\dontrun{
#Load the data
data("BinSpp")
#Use the function with the
Index <- GetQuadIndex(Stacklist = BinSpp, AmplFile = "https://raw.githubusercontent.com/derek-corcoran-barrios/QuadraticCostScripts/master/Two.txt")
}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
